mysite/requirements.txt
Django==4.0.2
djangorestframework==3.13.1
gunicorn==20.1.0
django-widget-tweaks==1.4.8
requests==2.27.1
joblib==1.4.2
django-crispy-forms==1.13.0
django-filter==21.1
django-cors-headers==3.10.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
gevent>=1.4
google-generativeai
python-dotenv
amadeus==7.1.0
whitenoise==6.3.0
django-use-email-as-username==1.4.0
geocoder==1.38.1
channels==3.0.5
channels_redis==4.2.0 
openpyxl==3.1.4
agora-token-builder==1.0.0
firebase-admin
google-cloud-storage
Pillow